backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
beautifulsoup4==4.12.2
Pillow==10.1.0
aiofiles==23.2.1
python-magic==0.4.27
requests==2.31.0
lxml==4.9.3

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
aiosqlite==0.19.0      # SQLite async

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0